abstract history table into common component
