+++ project/apps/web/components/collab/collab-table/CollabTableToolbar.tsx
@@ -27,14 +27,14 @@ interface DataTableToolbarProps<TData> {
   table: Table<TData>;
 }
 
-export function HistoryTableToolbar<TData>({
+export function CollabTableToolbar<TData>({
   table,
 }: DataTableToolbarProps<TData>) {
   const isFiltered = table.getState().columnFilters.length > 0;
 
   return (
     <div className="flex items-center justify-between">
-      <div className="flex flex-1 items-center space-x-2">
+      <div className="flex items-center flex-1 space-x-2">
         <Input
           placeholder="Filter questions..."
           value={
@@ -71,7 +71,7 @@ export function HistoryTableToolbar<TData>({
             className="h-8 px-2 lg:px-3"
           >
             Reset
-            <Cross2Icon className="ml-2 h-4 w-4" />
+            <Cross2Icon className="w-4 h-4 ml-2" />
           </Button>
         )}
       </div>
